Add -Doptimization=2 to Meson build option
This commit adds the -Doptimization=2 option to the Meson build command in order to set the optimization level to -O2 for the entire project. This optimization level is chosen to balance between compile time and runtime performance, improving the overall performance of the project.

This change ensures that the project is built with the desired optimization level, which is crucial for achieving optimal performance in production builds